+++ frontend/src/components/Service/ServiceInfo.vue
@@ -1,7 +1,8 @@
 <template v-if="service">
     <div class="col-12 card mb-4" style="min-height: 280px;" :class="{'offline-card': !service.online}">
         <div class="card-body p-3 p-md-1 pt-md-3 pb-md-1">
-            <h4 class="card-title mb-4"><router-link :to="serviceLink(service)">{{service.name}}</router-link>
+            <h4 class="card-title mb-4">
+                <router-link :to="serviceLink(service)">{{service.name}}</router-link>
                 <span class="badge float-right" :class="{'badge-success': service.online, 'badge-danger': !service.online}">
                     {{service.online ? "ONLINE" : "OFFLINE"}}
                 </span>
@@ -44,25 +45,23 @@
                 </div>
 
                     <div class="col-4">
-                        <button @click.prevent="Tab('incident')" class="btn btn-block btn-outline-secondary" :class="{'text-white btn-secondary': openTab==='incident'}" >Create Incident</button>
+                        <button @click.prevent="Tab('incident')" class="btn btn-block btn-outline-secondary incident" :class="{'text-white btn-secondary': openTab==='incident'}" >Incidents</button>
                     </div>
-                    <div class="col-4">
-                        <button @click.prevent="Tab('message')" class="btn btn-block btn-outline-secondary" :class="{'text-white btn-secondary': openTab==='message'}">Create Announcement</button>
-                    </div>
-                    <div class="col-4">
-                        <button @click.prevent="Tab('failures')" class="btn btn-block btn-outline-secondary" :disabled="service.stats.failures === 0" :class="{'text-white btn-secondary': openTab==='failures'}">
-                            Failures <span class="badge badge-danger float-right mt-1">{{service.stats.failures}}</span></button>
-                    </div>
+                <div class="col-4">
+                    <button @click.prevent="Tab('checkin')" class="btn btn-block btn-outline-secondary checkin" :class="{'text-white btn-secondary': openTab==='checkin'}" >Checkins</button>
+                </div>
+                <div class="col-4">
+                    <button @click.prevent="Tab('failures')" class="btn btn-block btn-outline-secondary failures" :disabled="service.stats.failures === 0" :class="{'text-white btn-secondary': openTab==='failures'}">
+                        Failures <span class="badge badge-danger float-right mt-1">{{service.stats.failures}}</span></button>
+                </div>
 
                 <div v-if="openTab === 'incident'" class="col-12 mt-4">
                     <FormIncident :service="service" />
                 </div>
 
-                <div v-if="openTab === 'message'" class="col-12 mt-4">
-                    <FormMessage :service="service"/>
-                </div>
-
                 <div v-if="openTab === 'failures'" class="col-12 mt-4">
+                    <button @click.prevent="deleteFailures" class="btn btn-block btn-outline-secondary delete_failures" :disabled="service.stats.failures === 0">Delete Failures</button>
+
                     <ServiceFailures :service="service"/>
                 </div>
 
@@ -121,6 +120,13 @@
           this.loaded = true
       },
       methods: {
+        async deleteFailures() {
+          const c = confirm('Are you sure you want to delete all failures?')
+          if (c) {
+            await Api.service_failures_delete(this.service)
+            this.service = await Api.service(this.service.id)
+          }
+        },
           Tab(name) {
               if (this.openTab === name) {
                   this.openTab = ''
@@ -129,7 +135,6 @@
               this.openTab=name;
           },
         sinceYesterday(data) {
-            window.console.log(data)
           let total = 0
           data.forEach((f) => {
             total += parseInt(f.y)