+++ frontend/src/forms/Login.vue
@@ -23,6 +23,10 @@
             </div>
         </div>
 
+        <a v-if="oauth.oauth_providers.split(',').includes('github')" class="btn btn-block btn-outline-dark" :href="`https://github.com/login/oauth/authorize?scope=user:email&client_id=${oauth.gh_client_id}`">Login with Github</a>
+        <a v-if="oauth.oauth_providers.split(',').includes('google')" class="btn btn-block btn-outline-secondary" :href="`https://accounts.google.com/signin/oauth?client_id=${oauth.google_client_id}&response_type=code&scope=${google_scope}&redirect_uri=${$store.getters.core.domain}/oauth/google`">Login with Google</a>
+        <a v-if="oauth.oauth_providers.split(',').includes('slack')" class="btn btn-block btn-outline-secondary" :href="`https://slack.com/oauth/v2/authorize?client_id=${oauth.slack_client_id}&team=${oauth.slack_team}&user_scope=${slack_scope}&redirect_uri=${$store.getters.core.domain}/oauth/slack`">Login with Slack</a>
+
     </form>
 </template>
 
@@ -31,6 +35,11 @@
 
   export default {
       name: 'FormLogin',
+    props: {
+        oauth: {
+          type: Object
+        }
+    },
       data() {
           return {
               username: "",
@@ -39,7 +48,8 @@
               loading: false,
               error: false,
               disabled: true,
-              ghLoginURL: ""
+            google_scope: "https%3A%2F%2Fwww.googleapis.com%2Fauth%2Fuserinfo.profile+https%3A%2F%2Fwww.googleapis.com%2Fauth%2Fuserinfo.email",
+            slack_scope: "identity.email,identity.basic"
           }
       },
       mounted() {