+++ frontend/vue.config.js
@@ -7,6 +7,10 @@ module.exports = {
       '/api': {
         logLevel: 'debug',
         target: 'http://0.0.0.0:8585'
+      },
+      '/oauth': {
+        logLevel: 'debug',
+        target: 'http://0.0.0.0:8585'
       }
     }
   }