+++ types/incidents/database_updates.go
@@ -2,22 +2,22 @@ package incidents
 
 func (i *Incident) Updates() []*IncidentUpdate {
 	var updates []*IncidentUpdate
-	db.Model(&IncidentUpdate{}).Where("incident = ?", i.Id).Find(&updates)
+	dbUpdate.Where("incident = ?", i.Id).Find(&updates)
 	i.AllUpdates = updates
 	return updates
 }
 
 func (i *IncidentUpdate) Create() error {
-	q := db.Create(i)
+	q := dbUpdate.Create(i)
 	return q.Error()
 }
 
 func (i *IncidentUpdate) Update() error {
-	q := db.Update(i)
+	q := dbUpdate.Update(i)
 	return q.Error()
 }
 
 func (i *IncidentUpdate) Delete() error {
-	q := db.Delete(i)
+	q := dbUpdate.Delete(i)
 	return q.Error()
 }